LayoutTests/media/modern-media-controls/time-label/time-label-expected.txt

Testing the TimeLabel class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS timeLabel.value is 0
PASS timeLabel.element.localName is "div"
PASS timeLabel.element.className is "time-label"
PASS style.fontFamily is "-apple-system-monospaced-numbers"
PASS style.fontSize is "12px"

NaNTimeLabel.element.textContent = --:--

PASS elaspedLabel is "Elapsed"
PASS remainingLabel is "Remaining"
PASS durationLabel is "Duration"
PASS accessibilityController.accessibleElementById('duration').role is "AXRole: AXStaticText"
PASS accessibilityController.accessibleElementById('remaining').role is "AXRole: AXStaticText"
PASS accessibilityController.accessibleElementById('elasped').role is "AXRole: AXStaticText"

elapsedTimeLabels[4].element.textContent = 00:01
remainingTimeLabels[4].element.textContent = -00:01
durationTimeLabels[4].element.textContent = 00:01

elapsedTimeLabels[5].element.textContent = 0:00:01
remainingTimeLabels[5].element.textContent = -0:00:01
durationTimeLabels[5].element.textContent = 0:00:01

elapsedTimeLabels[6].element.textContent = 00:00:01
remainingTimeLabels[6].element.textContent = -00:00:01
durationTimeLabels[6].element.textContent = 00:00:01

PASS successfullyParsed is true

TEST COMPLETE
